backend/audio_transcription/requirements.txt
fastapi
uvicorn[standard]
websockets
faster-whisper
numpy
scipy
noisereduce
webrtcvad-wheels
pydub
